requirements.txt
paramiko
pandas
matplotlib
pyyaml
requests